fixed game state test driver

+++ data_structures/synch.h
@@ -5,7 +5,7 @@
 
 UINT16 can_move_down(Model *model);
 
-UINT16 check_collision(const Matrix *board, const Tetromino *piece, UINT16 offset_x, UINT16 offset_y);
+UINT16 check_collision(const Matrix *board, const Tetromino *piece, int offset_x, int offset_y);
 
 void handle_tick (Model *model);
